feat(core.go): Add functions to write and read data
Move cassandra client related functions into core.go file.
